requirements.txt
Flask==2.3.3
requests==2.31.0
schedule==1.2.0
PyYAML==6.0.1
openai==1.3.5
requests[socks]==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
tweepy==4.14.0
